BMO-Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
llama-cpp-python==0.2.11
pydantic==2.5.0
psutil==5.9.0
python-multipart==0.0.6